+++ packaging/nix/default.nix
@@ -8,7 +8,7 @@
 
 stdenv.mkDerivation rec {
   pname = "retrace";
-  version = "2.13.0";
+  version = "2.14.0";
 
   src = ./.;